Move GeneratePartitionKey to main file
